Make Config an interface
